+++ demos/jest-express-demo/src/notebook.test.ts
@@ -238,7 +238,7 @@ class NoTransformResponseController {
      */
     @Get("/ping")
     ping() {
-        this.logger.info("Ping to services");
+        this.logger.debug("Ping to services");
         return {
             serviceA: this.serviceA.doSomething(),
             serviceN: this.serviceN.doSomethingElse(),
@@ -260,7 +260,7 @@ class NoTransformResponseController {
      */
     @Post("/default")
     default(@Body() user: UserModel) {
-        this.logger.info("Getting default");
+        this.logger.debug("Getting default");
         return handler(user);
     }
 
@@ -272,7 +272,7 @@ class NoTransformResponseController {
      */
     @Post("/noTransform", {transformRequest: false, transformResponse: false})
     noTransform(@Body() user: UserModel) {
-        this.logger.info("call noTransform");
+        this.logger.debug("call noTransform");
         return handler(user);
     }
 
@@ -284,7 +284,7 @@ class NoTransformResponseController {
      */
     @Post("/transformRequestOnly", {transformRequest: true, transformResponse: false})
     transformRequestOnly(@Body() user: UserModel) {
-        this.logger.info("call transformRequestOnly");
+        this.logger.debug("call transformRequestOnly");
         return handler(user);
     }
 
@@ -296,7 +296,7 @@ class NoTransformResponseController {
      */
     @Post("/transformResponseOnly", {transformRequest: false, transformResponse: true})
     transformResponseOnly(@Body() user: UserModel) {
-        this.logger.info("call transformResponseOnly");
+        this.logger.debug("call transformResponseOnly");
         return handler(user);
     }
 }