+++ demos/jest-express-demo/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @nodeboot/jest-express-demo
 
+## 1.5.0
+
+### Minor Changes
+
+-   Extend test framework to provide proper logging
+
 ## 1.4.0
 
 ### Minor Changes